services/saver/requirements.txt
pika==1.2.0
environ-config==21.2.0
backoff==1.11.1
